fix: Convierte ComprobanteId de nullable a int en CheckoutResultDto
El campo ComprobanteId en la respuesta de checkout requiere un
valor no nulo, pero comprobanteId se declaraba como int?.

+++ Services/Implementations/EstanciaService.cs
@@ -287,7 +287,7 @@ public class EstanciaService : IEstanciaService
                 TotalHabitacion = totalHabitacion,
                 TotalConsumos = totalConsumos,
                 TotalFinal = totalFinal,
-                ComprobanteId = comprobanteId
+                ComprobanteId = (int)comprobanteId
             };
         }
         catch
